requirements.txt
fastapi== 0.83.0
uvicorn==0.18.3
pytest==7.1.3
SPARQLWrapper==2.0.0
python-dotenv
jinja2